# optimize str.replace