@rust-log-analyzer 2024-09-23
The job **`mingw-check`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/10993045290/job/30518715907) [(plain)](https://github.com/rust-lang/rust/commit/176bcc0d01466713e8aa3bc6ad18d696653bfccc/checks/30518715907/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
#16 2.846 Building wheels for collected packages: reuse
#16 2.847   Building wheel for reuse (pyproject.toml): started
#16 3.092   Building wheel for reuse (pyproject.toml): finished with status 'done'
#16 3.092   Created wheel for reuse: filename=reuse-4.0.3-cp310-cp310-manylinux_2_35_x86_64.whl size=132715 sha256=dfa09868353292d98f811d3efdb0d54d07389e808efc71d68e3b93c514bf8bec
#16 3.093   Stored in directory: /tmp/pip-ephem-wheel-cache-vjaim2jj/wheels/3d/8d/0a/e0fc6aba4494b28a967ab5eaf951c121d9c677958714e34532
#16 3.095 Installing collected packages: boolean-py, binaryornot, tomlkit, reuse, python-debian, markupsafe, license-expression, jinja2, chardet, attrs
#16 3.484 Successfully installed attrs-23.2.0 binaryornot-0.4.4 boolean-py-4.0 chardet-5.2.0 jinja2-3.1.4 license-expression-30.3.0 markupsafe-2.1.5 python-debian-0.1.49 reuse-4.0.3 tomlkit-0.13.0
#16 3.484 WARNING: Running pip as the 'root' user can result in broken permissions and conflicting behaviour with the system package manager. It is recommended to use a virtual environment instead: https://pip.pypa.io/warnings/venv
#16 DONE 3.6s
---
    Checking rustc-demangle v0.1.24
error[E0433]: failed to resolve: use of undeclared type `Utf8Pattern`
   --> alloc/src/str.rs:273:18
    |
273 |             Some(Utf8Pattern::Slice([from_byte])) => Some(*from_byte),
    |                  ^^^^^^^^^^^ use of undeclared type `Utf8Pattern`
help: consider importing one of these enums
    |
10  + use crate::str::pattern::Utf8Pattern;
    |
    |
10  + use core::str::pattern::Utf8Pattern;
    |

error[E0433]: failed to resolve: use of undeclared type `Utf8Pattern`
   --> alloc/src/str.rs:274:18
    |
274 |             Some(Utf8Pattern::Char(c)) => c.as_ascii().map(|ascii_char| ascii_char.to_u8()),
    |                  ^^^^^^^^^^^ use of undeclared type `Utf8Pattern`
help: consider importing one of these enums
    |
10  + use crate::str::pattern::Utf8Pattern;
    |
```

</details>